Revert "Organize headers (#75)"
This reverts commit ba9af52f61c2a3ae60103f7bd658637db8f7c97d.
