+++ include/common.h
@@ -2,34 +2,9 @@
 #define _COMMON_H
 
 #include "ultra64.h"
-
-typedef u8 UNK_TYPE_8;
-typedef u16 UNK_TYPE_16;
-typedef u32 UNK_TYPE_32;
-typedef void UNK_PTR;
-
-#include "constants.h"
-#include "sys/fs.h"
-#include "sys/memory.h"
-#include "sys/thread.h"
-#include "sys/math.h"
-#include "sys/dll.h"
-#include "sys/debug.h"
-#include "sys/controller.h"
-#include "sys/gfx/gx.h"
-#include "sys/gfx/texture.h"
-#include "sys/gfx/animation.h"
-#include "sys/gfx/model.h"
-#include "sys/audio/audio.h"
-#include "sys/exception.h"
-#include "sys/crash.h"
-#include "sys/asset_thread.h"
-#include "game/gamebits.h"
-#include "game/actor/objdata.h"
-#include "game/actor/hitbox.h"
-#include "game/actor/actor.h"
-#include "game/actor/player/player.h"
-#include "variables.h"
+#include "common_structs.h"
 #include "functions.h"
+#include "variables.h"
+#include "constants.h"
 
-#endif //_COMMON_H
+#endif
