+++ include/constants.h
@@ -6,7 +6,15 @@
 #define HW_REG2(reg, type) *(volatile type *)(u32)(reg | 0xa0000000)
 #define ACCESS_1 *(s16 *)(u32)(0x1C000000 | 0xa0000000)
 #define ACCESS_2 *(s16 *)(u32)(0x1C000002 | 0xa0000000)
+#define EXPANSION_RAM_START 0x80400000
+#define EXPANSION_SIZE 0x00800000
 
-#define PLAYER_POSBUF_SIZE 60
+#define HEAP_AREA_00 0x8042C000
+#define HEAP_AREA_01 0x80245000
 
-#endif //_CONSTANTS_H
+#define IDLE_THREAD_SIZE (0x800 / sizeof(u64))
+#define MAIN_THREAD_SIZE 1024
+
+#define DEBUG_LOOP1 do {/* omitted debug code */} while(0)
+
+#endif
