--- include/game/actor/hitbox.h
@@ -1,9 +1,0 @@
-/** Actor hit detection system
- */
-
-typedef struct {
-/*0000*/    u8 unk_0x0[0x5a - 0x0];
-/*005A*/    u8 unk_0x5a;
-/*005B*/    u8 unk_0x5b[0x9f - 0x5b];
-/*009F*/    s8 unk_0x9f;
-} ActorObjhitInfo;
