--- include/game/gamebits.h
@@ -1,5 +1,0 @@
-/** GameBit (bittable) system
- */
-
-extern s32 gFile_BITTABLE;
-extern s16 gSizeBittable;
