--- include/sys/audio/audio.h
@@ -1,9 +1,0 @@
-/** Audio system
- */
-
-#ifndef _SYS_AUDIO_AUDIO_H
-#define _SYS_AUDIO_AUDIO_H
-
-extern u8 alSynFlag;
-
-#endif //_SYS_AUDIO_AUDIO_H
