--- include/sys/debug.h
@@ -1,8 +1,0 @@
-/** Misc debug stuff
- */
-#ifndef _SYS_DEBUG_H
-#define _SYS_DEBUG_H
-
-#define DEBUG_LOOP1 do {/* omitted debug code */} while(0)
-
-#endif //_SYS_DEBUG_H
