+++ src/crash.h
@@ -1,0 +1,10 @@
+#ifndef _CRASH_H_
+#define _CRASH_H_
+
+#include "ultra64.h"
+
+#define CRASH_THREAD_ID 0x64
+
+void start_crash_thread(OSSched* scheduler);
+
+#endif
