+++ symbol_addrs.txt
@@ -1217,12 +1217,3 @@ map_set_object_streaming_disabled = 0x800129D4; // type:func
 assetLoadThreadSendQueue = 0x800ACB48; // type:data
 assetLoadThreadRecvQueue = 0x800ACB98; // type:data
 assetLoadMsg = 0x800AE270; // type:data
-get_loaded_dlls = 0x8000BDE8; // type:func
-mainSetBits = 0x80014670; // type:func
-mainGetBit = 0x800147EC; // type:func
-mainDecrementBit = 0x800149AC; // type:func
-s_mm_audioheap = 0x80099228; // type:data
-doNothing_8001B4E4 = 0x8001B4E4; // type:func
-get_num_objects = 0x800212D8; // type:func
-update_obj_hitboxes = 0x80026C14; // type:func
-obj_do_hit_detection = 0x800271FC; // type:func
